AM: use stencil coordinates modulo grid size
Before this change in the case of zpad > L the stencil coordinates where out of
bounds, formally. In the normal case zpad == L this error does not occur. So we
will wrap the indices modulo grid size  to formally constrain the range. It has
no effect  in the normal case  and has a  meaningfull effect in the  case which
currently does not quite work anyway.

It is  not quite clear how  this case would  interact with the ghost  grid. The
hope is  with Vec  declared as  3d-periodic and  stencil size  1 access  to the
points offset by +1  and -1 in any of the  directions modulo periodicity should
be possible.