AM: print number of solvent molecules and hole integrals
A hole integral is  a negative of the integral g  - 1. It is
supposed to  reflect the number  of sites expelled  from the
neighbourhood of the solute/solvent. It is natural to expect
that, say, an HCl solute  immersed into the HCl solvent will
expell exactly one  solvent molecule. That is one  H and one
Cl site. Note the difference in BGY3D and BGY3DM numbers.